used float16 now

+++ distill_llama.py
@@ -54,7 +54,7 @@ def train(rank, world_size):
     load_in_4bit=True,
     bnb_4bit_use_double_quant=True,
     bnb_4bit_quant_type="nf4",
-    bnb_4bit_compute_dtype=torch.float32
+    bnb_4bit_compute_dtype=torch.float16
     )
     
     peft_config = LoraConfig(
